+++ app/components/user-menu/user-menu-directive.js
@@ -5,14 +5,13 @@
 angular.module('user-menu')
   .directive('userMenu', function () {
 
-  var link = function () {
-    if (window.Lizard) { window.Lizard.startPlugins() }; // jshint ignore:line
-  };
+    var link = function (scope, element, attrs) {
+      if (window.Lizard) { window.Lizard.startPlugins() }; // jshint ignore:line
+    };
 
-  return {
-    restrict: 'E',
-    replace: true,
-    link: link,
-    templateUrl: 'user-menu/user-menu.html'
-  };
+    return {
+      restrict: 'E',
+      link: link,
+      templateUrl: 'user-menu/user-menu.html'
+    };
   });