+++ test/karma.conf.dev.js
@@ -52,6 +52,7 @@ module.exports = function(config) {
       'app/components/timeseries/timeseries.js',
       'app/components/scenarios/scenarios.js',
       'app/components/user-menu/user-menu.js',
+      'app/components/favourites/favourites.js',
       'app/lizard-nxt.js',
       'test/mocks/**/*.js',
       'app/lib/**/*.js',