+++ test/karma.conf.js
@@ -52,6 +52,7 @@ module.exports = function(config) {
       'app/components/annotations/annotations.js',
       'app/components/user-menu/user-menu.js',
       'app/components/dashboard/dashboard.js',
+      'app/components/favourites/favourites.js',
       'app/lizard-nxt.js',
       'test/mocks/**/*.js',
       'app/lib/**/*.js',