+++ CHANGES.rst
@@ -41,6 +41,10 @@ Unreleased (2.6.2) (XXXX-XX-XX)
 
 - Fix data menu crosshair click is also toggles the layer visibility.
 
+- Fix translation of creation errors in annotations directive.
+
+- Add favourites.
+
 
 Release 2.7.1 (2016-1-29)
 ---------------------